requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
openai==1.3.7
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
google-generativeai==0.3.2
